core: Remove code duplication in tty_* data sources
SonarCloud link (there is no dedicated reference provided for this):
https://sonarcloud.io/component_measures?id=snoopy&metric=duplicated_lines_density&selected=snoopy%3Asrc%2Fdatasource%2Ftty_uid.c&view=list
